adjusted proofpoint.py to handle phish threats with Url
ProofPoint API has began returning Hostname only as a 'threat'. The Url.is_valid validation fails, as the hostname isn't a valid URL (lacks scheme, lacks path).

This revisions tries to add the threat as a Url; and if that fails will attempt to add it as a Hostname.

This gets the threat to the right object type; and improves ingestion of ProofPoint threats of type Phish.