Create circl_passive_ssl.py
Passive SSL plugin using Circl.lu api. I would create theses into certificate objects , but the api works on sha1 hashes and not actual data except in the case of the fetch certificate def. In any case to turn the output into a cert object would be a bit more work so for now this will just save the cert details in text nodes and in the raw data.